8268119: Rename copy_os_cpu.inline.hpp files to copy_os_cpu.hpp
