+++ src/hotspot/cpu/arm/copy_arm.hpp
@@ -25,12 +25,7 @@
 #ifndef CPU_ARM_COPY_ARM_HPP
 #define CPU_ARM_COPY_ARM_HPP
 
-#include "utilities/macros.hpp"
-
-// Inline functions for memory copy and fill.
-
-// Contains inline asm implementations
-#include OS_CPU_HEADER_INLINE(copy)
+#include OS_CPU_HEADER(copy)
 
 static void pd_fill_to_words(HeapWord* tohw, size_t count, juint value) {
   juint* to = (juint*)tohw;
