+++ src/hotspot/cpu/x86/copy_x86.hpp
@@ -25,10 +25,7 @@
 #ifndef CPU_X86_COPY_X86_HPP
 #define CPU_X86_COPY_X86_HPP
 
-// Inline functions for memory copy and fill.
-
-// Contains inline asm implementations
-#include OS_CPU_HEADER_INLINE(copy)
+#include OS_CPU_HEADER(copy)
 
 static void pd_fill_to_words(HeapWord* tohw, size_t count, juint value) {
 #ifdef AMD64
