+++ src/hotspot/os_cpu/linux_arm/copy_linux_arm.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Microsoft Corporation. All rights reserved.
+ * Copyright (c) 2008, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,53 +22,23 @@
  *
  */
 
-#ifndef OS_CPU_WINDOWS_AARCH64_COPY_WINDOWS_AARCH64_INLINE_HPP
-#define OS_CPU_WINDOWS_AARCH64_COPY_WINDOWS_AARCH64_INLINE_HPP
-
-#include <string.h>
+#ifndef OS_CPU_LINUX_ARM_COPY_LINUX_ARM_HPP
+#define OS_CPU_LINUX_ARM_COPY_LINUX_ARM_HPP
 
 static void pd_conjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
-  (void)memmove(to, from, count * HeapWordSize);
+  _Copy_conjoint_words(from, to, count * HeapWordSize);
 }
 
 static void pd_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
-  switch (count) {
-  case 8:  to[7] = from[7];
-  case 7:  to[6] = from[6];
-  case 6:  to[5] = from[5];
-  case 5:  to[4] = from[4];
-  case 4:  to[3] = from[3];
-  case 3:  to[2] = from[2];
-  case 2:  to[1] = from[1];
-  case 1:  to[0] = from[0];
-  case 0:  break;
-  default:
-    (void)memcpy(to, from, count * HeapWordSize);
-    break;
-  }
+  _Copy_disjoint_words(from, to, count * HeapWordSize);
 }
 
 static void pd_disjoint_words_atomic(const HeapWord* from, HeapWord* to, size_t count) {
-  switch (count) {
-  case 8:  to[7] = from[7];
-  case 7:  to[6] = from[6];
-  case 6:  to[5] = from[5];
-  case 5:  to[4] = from[4];
-  case 4:  to[3] = from[3];
-  case 3:  to[2] = from[2];
-  case 2:  to[1] = from[1];
-  case 1:  to[0] = from[0];
-  case 0:  break;
-  default: while (count-- > 0) {
-             *to++ = *from++;
-           }
-           break;
-  }
+  pd_disjoint_words(from, to, count);
 }
 
 static void pd_aligned_conjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
- // pd_conjoint_words(from, to, count);
-  (void)memmove(to, from, count * HeapWordSize);
+  pd_conjoint_words(from, to, count);
 }
 
 static void pd_aligned_disjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
@@ -76,7 +46,7 @@ static void pd_aligned_disjoint_words(const HeapWord* from, HeapWord* to, size_t
 }
 
 static void pd_conjoint_bytes(const void* from, void* to, size_t count) {
-  (void)memmove(to, from, count);
+  memmove(to, from, count);
 }
 
 static void pd_conjoint_bytes_atomic(const void* from, void* to, size_t count) {
@@ -84,59 +54,26 @@ static void pd_conjoint_bytes_atomic(const void* from, void* to, size_t count) {
 }
 
 static void pd_conjoint_jshorts_atomic(const jshort* from, jshort* to, size_t count) {
-  if (from > to) {
-    while (count-- > 0) {
-      // Copy forwards
-      *to++ = *from++;
-    }
-  } else {
-    from += count - 1;
-    to   += count - 1;
-    while (count-- > 0) {
-      // Copy backwards
-      *to-- = *from--;
-    }
-  }
+  _Copy_conjoint_jshorts_atomic(from, to, count * BytesPerShort);
 }
 
 static void pd_conjoint_jints_atomic(const jint* from, jint* to, size_t count) {
-  if (from > to) {
-    while (count-- > 0) {
-      // Copy forwards
-      *to++ = *from++;
-    }
-  } else {
-    from += count - 1;
-    to   += count - 1;
-    while (count-- > 0) {
-      // Copy backwards
-      *to-- = *from--;
-    }
-  }
+  assert(HeapWordSize == BytesPerInt, "heapwords and jints must be the same size");
+  // pd_conjoint_words is word-atomic in this implementation.
+  pd_conjoint_words((const HeapWord*)from, (HeapWord*)to, count);
 }
 
 static void pd_conjoint_jlongs_atomic(const jlong* from, jlong* to, size_t count) {
-  pd_conjoint_oops_atomic((const oop*)from, (oop*)to, count);
+  _Copy_conjoint_jlongs_atomic(from, to, count * BytesPerLong);
 }
 
 static void pd_conjoint_oops_atomic(const oop* from, oop* to, size_t count) {
- if (from > to) {
-    while (count-- > 0) {
-      // Copy forwards
-      *to++ = *from++;
-    }
-  } else {
-    from += count - 1;
-    to   += count - 1;
-    while (count-- > 0) {
-      // Copy backwards
-      *to-- = *from--;
-    }
-  }
+  assert(BytesPerHeapOop == BytesPerInt, "32-bit architecture");
+  pd_conjoint_jints_atomic((const jint*)from, (jint*)to, count);
 }
 
 static void pd_arrayof_conjoint_bytes(const HeapWord* from, HeapWord* to, size_t count) {
-  pd_conjoint_bytes_atomic(from, to, count);
+  pd_conjoint_bytes_atomic((const void*)from, (void*)to, count);
 }
 
 static void pd_arrayof_conjoint_jshorts(const HeapWord* from, HeapWord* to, size_t count) {
@@ -152,7 +89,7 @@ static void pd_arrayof_conjoint_jlongs(const HeapWord* from, HeapWord* to, size_
 }
 
 static void pd_arrayof_conjoint_oops(const HeapWord* from, HeapWord* to, size_t count) {
- pd_conjoint_oops_atomic((const oop*)from, (oop*)to, count);
+  pd_conjoint_oops_atomic((const oop*)from, (oop*)to, count);
 }
 
-#endif // OS_CPU_WINDOWS_AARCH64_COPY_WINDOWS_AARCH64_INLINE_HPP
+#endif // OS_CPU_LINUX_ARM_COPY_LINUX_ARM_HPP
