+++ src/hotspot/os_cpu/linux_x86/copy_linux_x86.hpp
@@ -22,8 +22,8 @@
  *
  */
 
-#ifndef OS_CPU_LINUX_X86_COPY_LINUX_X86_INLINE_HPP
-#define OS_CPU_LINUX_X86_COPY_LINUX_X86_INLINE_HPP
+#ifndef OS_CPU_LINUX_X86_COPY_LINUX_X86_HPP
+#define OS_CPU_LINUX_X86_COPY_LINUX_X86_HPP
 
 static void pd_conjoint_words(const HeapWord* from, HeapWord* to, size_t count) {
 #ifdef AMD64
@@ -306,4 +306,4 @@ static void pd_arrayof_conjoint_oops(const HeapWord* from, HeapWord* to, size_t 
 #endif // AMD64
 }
 
-#endif // OS_CPU_LINUX_X86_COPY_LINUX_X86_INLINE_HPP
+#endif // OS_CPU_LINUX_X86_COPY_LINUX_X86_HPP
